+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This document provides a log of changes. It is currently maintained by hand.
 
 ## (master)
 
+## v0.3
+
 - Fixed index.d.ts declarations missing from the published package...
 - Introduced the `quickstrap` function
 - Introduced `postSetup` module method